Switched to new Type class, prepared for Plic Type Package utilization.
* rapicorn-core/types.hh, types.cc: replaced the old Enumeration and Type classes
with the new Typ2 class which got renamed to class Type. The new class provides
a simpler API and needs only one class to represent all types. The best storage
type identifiers to mediate between semantic types as provided by plic and
implementation types like Value and Array are yet to be determined.

* rapicorn-core/types.cc: include types-ptpp.cc to parse PLIC type packages.

* rapicorn-core/types-ptpp.cc: PLIC TypePackage Parser, checked in because plic
is a devel-only dependency.

* rapicorn-core/values.hh, values.cc: use the new Type class. Alias some storage
types, which boil down to the same implementation mechanisms. renamed real().

* rapicorn-core/tests/mod1tst.cc: disabled all test, broken by the new Type class.

* rapicorn-core/tests/modval.cc: disabled all test, all are broken currently.

* configure.in: check for plic, provide $(PLIC).

* rapicorn-core/Makefile.am: generate types-ptpp.cc with plic.