Error out if ReferenceCountImpl objects are allocated on the stack.
* rapicornutils.hh: check that reference counted instances are not allocated
on the stack in ReferenceCountImpl::ReferenceCountImpl() unless a specific
magic has been passed on construction (because the need to work around this
check is expected to be exceptionally rare).

* rapicornutils.cc: error out in ReferenceCountImpl::stackcheck() if the
instance is allocated on the stack (probability check, based on 'this' being
in page_size proximity of the stack pointer).
slight fixup to pointer address modulo in malloc_aligned().

* rapicornsignal.hh: work around the stack-allocation-check in
ReferenceCountImpl for SignalBase::EmbeddedLink which has very controlled
allocation and reference counting semantics. Needed because objects with
signals may be allocated on the stack.