UI: only consider logical positions for logical focus movements
Tab and Shift+Tab are used for "logical" focus movement, i.e. move focus
through the order in which the widgets are layed out and allocated. This
works *regardless* of scroll positions or temporary invisible regions.
As such, we need to pick up the real allocation and not clipped or other-
wise constrained areas to determine the focus chain.

Signed-off-by: Tim Janik <timj@gnu.org>